miss

+++ common/csv_loader.py
@@ -21,7 +21,7 @@ def get_featured_dtypes():
         'os': 'uint16',
         'channel': 'uint16',
         'is_attributed': 'uint8',
-        'click_id': 'uint32',
+        'click_id': 'float32',
         'hour': 'uint16',
         'app_count': 'uint32',
         'os_count': 'uint32',
